// LungLife/lunglife_backend/src/domain/interfaces/profile.interface.ts
export enum UserRole {
  PATIENT = 'patient',
  HEALTH_PROFESSIONAL = 'health_professional',
  ADMIN = 'admin',
  RESEARCHER = 'researcher'
}